requirements.txt
pandas
pyarrow
pytest
gunicorn==20.1.0
Flask=2.1.0
google-cloud-bigquery==2.34.4
